fix: split openapi contract ref helpers for backend-size limit
Move assertOpenAPI3RefPropertyDescription and resolveOpenAPI3ReferencedPropertySchema into a sibling test file so openapi_contract_common_test.go stays within the 1000-line backend-size gate (addresses PR #1143 BLOCKING feedback).
