+++ pkg/transports/http/contracttests/openapi_contract_common_test.go
@@ -405,32 +405,6 @@ func assertOpenAPI3ArrayPropertyDescription(t *testing.T, schema *openapi3.Schem
 	return property.Items.Value
 }
 
-func assertOpenAPI3RefPropertyDescription(t *testing.T, schema *openapi3.Schema, schemaName string, propertyName string) *openapi3.Schema {
-	t.Helper()
-	property, ok := schema.Properties[propertyName]
-	if !ok || property == nil || property.Value == nil {
-		t.Fatalf("%s.properties.%s is missing", schemaName, propertyName)
-	}
-	assertOpenAPI3Description(t, schemaName+".properties."+propertyName, property.Value.Description)
-	return resolveOpenAPI3ReferencedPropertySchema(t, property.Value, schemaName+".properties."+propertyName)
-}
-
-func resolveOpenAPI3ReferencedPropertySchema(t *testing.T, schema *openapi3.Schema, path string) *openapi3.Schema {
-	t.Helper()
-	if schema == nil {
-		t.Fatalf("%s schema is missing", path)
-	}
-	if len(schema.AllOf) == 1 {
-		if schema.AllOf[0].Value != nil {
-			return schema.AllOf[0].Value
-		}
-		if schema.AllOf[0].Ref != "" {
-			t.Fatalf("%s allOf[0] ref %q is not resolved", path, schema.AllOf[0].Ref)
-		}
-	}
-	return schema
-}
-
 func assertOpenAPI3PropertyRef(t *testing.T, schema *openapi3.Schema, schemaName string, propertyName string, wantRef string) {
 	t.Helper()
 	property, ok := schema.Properties[propertyName]
